MTSApi/requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.8.0
asyncio==3.4.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
DAWG-Python==0.7.2
docopt==0.6.2
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
idna==3.10
intervaltree==3.1.0
ipymarkup==0.9.0
Jinja2==3.1.5
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
navec==0.10.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
pydantic==2.10.6
pydantic_core==2.27.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
razdel==0.5.0
regex==2024.11.6
requests==2.32.3
slovnet==0.6.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.45.3
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yargy==0.16.0
